refactor: split ollama transport and parsing
