+++ src/ollama/mod.rs
@@ -1,2 +1,4 @@
 pub mod client;
+pub(crate) mod parsing;
+pub(crate) mod transport;
 pub mod xml_fallback;
